`internal/config/prometheus.go`
